pull out BooleanFunctionEqualsElimination tests
